YapDatabase: adapt for Xcode 26.5.
Drop unused #import <netinet6/in6.h> from YapReachability.m; the
header is now flagged as a private module-internal include under
Xcode 26.5.

Cast YapDatabaseConnectionFlushMemoryFlags_Extension_State to
NSUInteger at the bitwise sites in YapDatabaseConnection.m and
YapDatabaseViewConnection.m to silence -Wenum-enum-conversion, now
an error under Xcode 26.5.

+++ YapDatabase/Extensions/View/YapDatabaseViewConnection.m
@@ -99,7 +99,7 @@
 		[self _flushStatements];
 	}
 	
-	if (flags & YapDatabaseConnectionFlushMemoryFlags_Extension_State)
+	if (flags & (NSUInteger)YapDatabaseConnectionFlushMemoryFlags_Extension_State)
 	{
 		state = nil;
 	}
