+++ YapDatabase/YapDatabaseConnection.m
@@ -2586,7 +2586,7 @@ static int connectionBusyHandler(void *ptr, int count)
 		if (!changesets) // we could not retrieve changeset due to a change from another process.
 		{
 			NSUInteger flags = YapDatabaseConnectionFlushMemoryFlags_Caches |
-			                   YapDatabaseConnectionFlushMemoryFlags_Extension_State;
+			                   (NSUInteger)YapDatabaseConnectionFlushMemoryFlags_Extension_State;
 			
 			[self _flushMemoryWithFlags:flags];
 			snapshot = dbSnapshot;
@@ -2918,7 +2918,7 @@ static int connectionBusyHandler(void *ptr, int count)
 		if (!changesets) // we could not retrieve changeset due to a change from another process.
 		{
 			NSUInteger flags = YapDatabaseConnectionFlushMemoryFlags_Caches |
-			                   YapDatabaseConnectionFlushMemoryFlags_Extension_State;
+			                   (NSUInteger)YapDatabaseConnectionFlushMemoryFlags_Extension_State;
 			
 			[self _flushMemoryWithFlags:flags];
 			snapshot = dbSnapshot;
@@ -4188,7 +4188,7 @@ static int connectionBusyHandler(void *ptr, int count)
 	if (changeset_modifiedExternally)
 	{
 		NSUInteger flags = YapDatabaseConnectionFlushMemoryFlags_Caches |
-		                   YapDatabaseConnectionFlushMemoryFlags_Extension_State;
+		                   (NSUInteger)YapDatabaseConnectionFlushMemoryFlags_Extension_State;
 		
 		[self _flushMemoryWithFlags:flags];
 	}
@@ -4542,7 +4542,7 @@ static int connectionBusyHandler(void *ptr, int count)
 		snapshot = changesetSnapshot;
 		
 		NSUInteger flags = YapDatabaseConnectionFlushMemoryFlags_Caches |
-		                   YapDatabaseConnectionFlushMemoryFlags_Extension_State;
+		                   (NSUInteger)YapDatabaseConnectionFlushMemoryFlags_Extension_State;
 		
 		[self _flushMemoryWithFlags:flags];
 		[self processChangeset:changeset];
